# dashboard/src/api/openapi.yaml
openapi: 3.0.3
info:
  title: MONOLITH OS API
  description: |
    RESTful API for MONOLITH OS - An AI-powered enterprise operating system that orchestrates
    autonomous agents for business operations, decision-making, and workflow management.

    ## Features
    - **Task Management**: Create, update, and track tasks across the organization
    - **Decision Logging**: Record and audit business decisions with full traceability
    - **Workflow Orchestration**: Manage multi-step business processes
    - **Agent Management**: Monitor and control AI agents across different roles

    ## Authentication
    All endpoints require Bearer token authentication. Obtain a token via `/auth/login`.

    ## Rate Limiting
    - Standard endpoints: 100 requests/minute
    - Bulk operations: 10 requests/minute

    ## Versioning
    The API uses URL-based versioning. Current version: v1
  version: 1.0.0
  contact:
    name: MONOLITH OS Support
    email: support@monolith-os.local
  license:
    name: Proprietary
    url: https://monolith-os.local/license
servers:
  - url: http://localhost:3000/api
    description: Development server
  - url: https://api.monolith-os.local/api
    description: Production server
...
security:
  - BearerAuth: []
paths:
...
  /v1/escalations:
    get:
      tags:
        - Escalations
      summary: List escalations
      description: Retrieve pending escalations requiring CEO/human decision
      responses:
        '200':
          description: List of escalations
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EscalationListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agents/escalations/{escalationId}/resolve:
    post:
      tags:
        - Escalations
      summary: Resolve an escalation
      description: Resolve a pending escalation with a decision
      parameters:
        - name: escalationId
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - decision
              properties:
                decision:
                  type: string
                  description: The decision made
                notes:
                  type: string
                  description: Additional notes
      responses:
        '200':
          description: Escalation resolved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'
...
  /v1/agents:
    get:
      tags:
        - Agents
      summary: List agents
      description: Retrieve all registered agents
      parameters:
        - name: status
          in: query
          description: Filter by agent status
          schema:
            type: string
            enum: [active, idle, disabled, error]
        - name: enabled
          in: query
          description: Filter by enabled state
          schema:
            type: boolean
      responses:
        '200':
          description: List of agents
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentListResponse'
        '500':
          $ref: '#/components/responses/InternalError'
  /v1/agents/status:
    get:
      tags:
        - Agents
      summary: Get agent system status
      description: Retrieve overall agent system status including orchestrator state
      responses:
        '200':
          description: Agent system status
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AgentStatusResponse'
        '500':
          $ref: '#/components/responses/InternalError'
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /auth/login
...
  schemas:
...
    Escalation:
      type: object
      properties:
        id:
          type: string
        taskId:
          type: string
        reason:
          type: string
        priority:
          type: string
        createdAt:
          type: string
          format: date-time
        status:
          type: string
          enum: [pending, resolved]
...
    Agent:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        role:
          type: string
        status:
          type: string
          enum: [active, idle, disabled, error]
        enabled:
          type: boolean
        capabilities:
          type: array
          items:
            type: string
        metrics:
          type: object
          properties:
            tasksProcessed:
              type: integer
            tasksCompleted:
              type: integer
            tasksFailed:
              type: integer
            averageProcessingTime:
              type: number
            lastActiveAt:
              type: string
              format: date-time
              nullable: true
        config:
          type: object
          properties:
            maxConcurrentTasks:
              type: integer
            timeout:
              type: integer
            retryAttempts:
              type: integer
            priority:
              type: integer
...
    SuccessResponse:
      type: object
      properties:
        success:
          type: boolean
          example: true
        message:
          type: string
    ErrorResponse:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
            field:
              type: string
...
    EscalationListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Escalation'
        total:
          type: integer
...
    AgentListResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: array
          items:
            $ref: '#/components/schemas/Agent'
        total:
          type: integer
        systemStatus:
          type: object
          properties:
            orchestratorRunning:
              type: boolean
            mode:
              type: string
            queuedTasks:
              type: integer
            inProgress:
              type: integer
    AgentStatusResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
          properties:
            system:
              type: object
              properties:
                available:
                  type: boolean
                mode:
                  type: string
                orchestratorRunning:
                  type: boolean
            agents:
              type: object
              properties:
                total:
                  type: integer
                active:
                  type: integer
                disabled:
                  type: integer
            queue:
              type: object
              properties:
                pending:
                  type: integer
                inProgress:
                  type: integer
                completed:
                  type: integer
            escalations:
              type: object
              properties:
                pendingCEODecisions:
                  type: integer
            timestamp:
              type: string
              format: date-time
...
  responses:
    BadRequest:
      description: Bad request - validation error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: VALIDATION_ERROR
              message: Invalid input
              field: content
...
    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: NOT_FOUND
              message: Resource not found
    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            success: false
            error:
              code: INTERNAL_ERROR
              message: An unexpected error occurred